converted block speed into a double to modify speed more gradually. Made the speed accelerate/decelerate with proporiton to the square of the speed, just like a boat does.

+++ externalVariables.h
@@ -1,4 +1,4 @@
 //declared variables (no memory is allocated)
 extern volatile int pixel_buffer_start; 
-extern int dx_Boat;
-extern int dy_Boat;
+extern double dx_Boat;
+extern double dy_Boat;
